Fixed bugs and implemented registers
Fixed the bugs with the buscontroller, and implemented the skeleton for the registers.

+++ src/bus/device/ram.cpp
@@ -1,7 +1,7 @@
 #include "ram.h"
 #include <iostream>
 
-Ram::Ram(int ramSize) {
+Ram::Ram(uint16_t ramSize) {
     ram = new uint8_t[ramSize](); 
 }
